requirements.txt
flask==2.3.4
mysql-connector-python==8.1.0
pandas==2.1.1
matplotlib==3.8.0
